Added better check on self-signed certs. for #14

+++ start-mailserver.sh
@@ -31,16 +31,25 @@ postmap /etc/postfix/vmailbox
 postmap /etc/postfix/virtual
 cat /tmp/vhost.tmp | sort | uniq >> /etc/postfix/vhost && rm /tmp/vhost.tmp
 
-# Adding SSL certificate if provided in 'postfix/ssl' folder
-if [ -e "/tmp/postfix/ssl/$(hostname)-cert.pem" ]; then
+# Adding self-signed SSL certificate if provided in 'postfix/ssl' folder
+if [ -e "/tmp/postfix/ssl/$(hostname)-cert.pem" ] \
+&& [ -e "/tmp/postfix/ssl/$(hostname)-key.pem"  ] \
+&& [ -e "/tmp/postfix/ssl/$(hostname)-combined.pem" ] \
+&& [ -e "/tmp/postfix/ssl/demoCA/cacert.pem" ]; then
   echo "Adding $(hostname) SSL certificate"
-  cp -r /tmp/postfix/ssl /etc/postfix/ssl 
+  mkdir -p /etc/postfix/ssl
+  cp /tmp/postfix/ssl/$(hostname)-cert.pem /etc/postfix/ssl
+  cp /tmp/postfix/ssl/$(hostname)-key.pem /etc/postfix/ssl
+  cp /tmp/postfix/ssl/$(hostname)-combined.pem /etc/postfix/ssl
+  cp /tmp/postfix/ssl/demoCA/cacert.pem /etc/postfix/ssl
+
   # Postfix configuration
   sed -i -r 's/smtpd_tls_cert_file=\/etc\/ssl\/certs\/ssl-cert-snakeoil.pem/smtpd_tls_cert_file=\/etc\/postfix\/ssl\/'$(hostname)'-cert.pem/g' /etc/postfix/main.cf
   sed -i -r 's/smtpd_tls_key_file=\/etc\/ssl\/private\/ssl-cert-snakeoil.key/smtpd_tls_key_file=\/etc\/postfix\/ssl\/'$(hostname)'-key.pem/g' /etc/postfix/main.cf
   sed -i -r 's/#smtpd_tls_CAfile=/smtpd_tls_CAfile=\/etc\/postfix\/ssl\/demoCA\/cacert.pem/g' /etc/postfix/main.cf
   sed -i -r 's/#smtp_tls_CAfile=/smtp_tls_CAfile=\/etc\/postfix\/ssl\/demoCA\/cacert.pem/g' /etc/postfix/main.cf
   ln -s /etc/postfix/ssl/demoCA/cacert.pem /etc/ssl/certs/cacert-$(hostname).pem
+
   # Courier configuration
   sed -i -r 's/TLS_CERTFILE=\/etc\/courier\/imapd.pem/TLS_CERTFILE=\/etc\/postfix\/ssl\/'$(hostname)'-combined.pem/g' /etc/courier/imapd-ssl
 fi
